Add files via upload
This is the revamped code with the goals achieved. New version includes a file system, after processing the .xls from the ecoplate outfile insert them into the Input Folder. When you run the program they will get transformed and the transformation will be saved into the Output Folder. Those Files will be moved to the input_processed folder. I have included a sample set of 15 as a demo, also I added the ability to filter out files that don't include .xlsx as they will not work with the function. Only xlsx files specifically copied in the format I describe in the readme will be able to be transformed.