Add files via upload
This is the whole system of files including a demo of 15 files. Simply run the program and .xlsx files from the input folder will be transformed into a human readable format and saved in the output folder. The files from the input folder will then be moved to the input_processed folder after they have passed through the transformation. Remember this will only work for very specific .xlsx files that are output from an EcoPlate machine.